requirements.txt
colorzero==2.0
gpiozero==1.6.2
paho-mqtt==1.5.1
pkg-resources==0.0.0
pms5003==0.0.5
pyserial==3.5
RPi.bme280==0.2.3
RPi.GPIO==0.7.0
smbus2==0.4.1
spidev==3.5
